inprogress shadow test

+++ core/include/NinthEngine/Utils/MathUtils.hpp
@@ -15,7 +15,7 @@ namespace NinthEngine {
 typedef glm::vec4 Color;
 
 struct Viewport {
-	float x, y, width, height;
+	int x, y, width, height;
 };
 
 struct AABB {
